Added Proposal
This was our original proposal for the project. Our feedback was "Question 7 seems a bit open ended. It would be a nice big question that you place 3-6, 8, 9 under but restricted to final 4. I would want to know if elite recruits were known to be elite prior to the tournament or in hindsight."